+++ cm/cmind/repos.py
@@ -612,9 +612,27 @@ class Repos:
                 if console:
                     print ('  Deleting repository content ...')
 
-                shutil.rmtree(path_to_repo)
+                shutil.rmtree(path_to_repo, onerror=rm_read_only)
             else:
                 if console:
                     print ('  CM repository was unregistered from CM but its content was not deleted ...')
 
         return {'return':0}
+
+##############################################################################
+def rm_read_only(f, p, e):
+    """
+    Internal aux function to remove files and dirs even if read only
+    particularly on Windows
+    """
+
+    import os
+    import stat
+    import errno
+
+    ex = e[1]
+
+    os.chmod(p, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
+    f(p)
+
+    return