+++ cm/CHANGES.md
@@ -3,6 +3,9 @@
    - added support to show extra info about CM Git repos: "cm show repo"
    - added explicit support for Private Access Token (PAT) when pulling private CM repos:
        cm pull repo ctuning@mlcommons-ck-reproduce-inference-v4.0 --pat={GITHUB PAT}
+   - added support to remove CM repositories via "cm rm repo" even if read only
+       (required for Windows)
+
 
 ## V2.0.1
    - added setuptools as dependency to detect package versions